Refactor out Option class
