+++ src/models/Option.js
@@ -1,0 +1,12 @@
+class Option {
+  constructor(optionName) {
+    this.optionName = optionName;
+    this.isSelected = false;
+  }
+
+  setSelected(isSelected) {
+    this.isSelected = isSelected;
+  }
+}
+
+export default Option;
